pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dpdumper"
version = "0.0.2"
description = "Tool to use the dupico as a dumping device"
authors = [
   { name = "Fabio Battaglia", email = "hkzlabnet@gmail.com" }
]
keywords = ["dupico", "DuPAL", "dumping"]
readme = "README.md"
license = { text = "CC BY-SA 4.0 DEED" }
classifiers = [
    "Programming Language :: Python :: 3"
]
requires-python = ">=3.12"
dependencies = [
    "pyserial ~= 3.5",
    "dupicolib ~= 0.1.0"
]

[project.scripts]
dpdumper = 'dpdumper.frontend:cli'

[tool.setuptools]
py-modules = [ "__init__" ]

[project.urls]
repository = "https://github.com/hkzlab/dupico_dumper"